plug in receiver to save all messages.
The current model is for now going to expect app-level processing to happen on an individual basic for each app, rather than as each SMS is received. This might be a longer term change or might be just for a few commits. This is a quick and dirty way to make sure things get saved without much data loss, however, which seems like a win.